dedeup send traffic and decouple traffic shifting
