+++ pkg/test/framework/features/features.yaml
@@ -35,6 +35,7 @@ features:
   # features relating to controlling the traffic of the service mesh.
   traffic:
     reachability:
+    shifting:
     routing:
     short-circuiting:
     mirroring:
