+++ tests/integration/pilot/vm/util.go
@@ -16,6 +16,17 @@
 
 package vm
 
+import (
+	"fmt"
+	"math"
+	"strings"
+	"testing"
+	"time"
+
+	"istio.io/istio/pkg/test/framework/components/echo"
+	"istio.io/istio/pkg/test/util/retry"
+)
+
 const (
 	DefaultVMImage = "app_sidecar_ubuntu_bionic"
 )
@@ -23,4 +34,42 @@ const (
 func GetSupportedOSVersion() []string {
 	return []string{"app_sidecar_ubuntu_xenial", "app_sidecar_ubuntu_focal", "app_sidecar_ubuntu_bionic",
 		"app_sidecar_debian_9", "app_sidecar_debian_10"}
+}
+
+func SendTraffic(t *testing.T, batchSize int, from, to echo.Instance, hosts []string, weight []int32, errorThreshold float64) {
+	t.Helper()
+	// Send `batchSize` requests and ensure they are distributed as expected.
+	retry.UntilSuccessOrFail(t, func() error {
+		resp, err := from.Call(echo.CallOptions{
+			Target:   to,
+			PortName: "http",
+			Count:    batchSize,
+		})
+		if err != nil {
+			return fmt.Errorf("error during call: %v", err)
+		}
+		var totalRequests int
+		hitCount := map[string]int{}
+		for _, r := range resp {
+			for _, h := range hosts {
+				if strings.HasPrefix(r.Hostname, h+"-") {
+					hitCount[h]++
+					totalRequests++
+					break
+				}
+			}
+		}
+
+		for i, v := range hosts {
+			percentOfTrafficToHost := float64(hitCount[v]) * 100.0 / float64(totalRequests)
+			deltaFromExpected := math.Abs(float64(weight[i]) - percentOfTrafficToHost)
+			if errorThreshold-deltaFromExpected < 0 {
+				return fmt.Errorf("unexpected traffic weight for host %v. Expected %d%%, got %g%% (thresold: %g%%)",
+					v, weight[i], percentOfTrafficToHost, errorThreshold)
+			}
+			t.Logf("Got expected traffic weight for host %v. Expected %d%%, got %g%% (thresold: %g%%)",
+				v, weight[i], percentOfTrafficToHost, errorThreshold)
+		}
+		return nil
+	}, retry.Delay(time.Second))
 }
